requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
selenium==4.16.0
psycopg2==2.9.9
marshmallow==3.20.1
python-dotenv==1.0.1
schedule==1.2.2
requests==2.32.3